opencv/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
opencv-python-headless==4.10.0.84
numpy==2.1.1
python-multipart==0.0.9

# OCR
pytesseract==0.3.13
Pillow==10.4.0
